Tidy up presenters and helpers
Includes:

- moving search helper functions into their own file
- giving presenters their own module
- spliting search filters methods into their own class
- moving search filters helper into presenters
- adding comments to all the new code
